Rm output folder only if exists.

+++ web_shell/local_cloud/local_cloud.rb
@@ -53,7 +53,9 @@ end
 CC.logger.info("\n\n\n\n\n")
 
 # clear output
-FileUtils.rm_r("../agents_generator/cloud_agents_generated")
+if File.exists?("../agents_generator/cloud_agents_generated")
+  FileUtils.rm_r("../agents_generator/cloud_agents_generated")
+end
 FileUtils.mkdir_p("#{source_path}/cloud_agents_generated")
 
 #progen generation
